Fix Import Syntax
When we were trying to import the `FormattedMessage` component as though it were a default export (as shown in the first examples of the docs), we got errors. We figured out that the later examples are correct and that the component needs to be imported by name. This change updates the docs so that they're consistent and won't confuse others down the road.